==> 0_taudem_dem_prep.Rmd <==
---
title: "Methods for: Identifying geomorphic process domains in synthetic landscapes of West Virginia, USA"
author:
- "Dr. Matthew Ross and Dr. Kristin Jaeger"
- 'matt.ross(at)colostate.edu'
date: '`r format(Sys.Date(), "%B %d, %Y")`'
site: bookdown::bookdown_site
documentclass: book
output:
  bookdown::gitbook: default
  #bookdown::pdf_book: default
# bibliography: 'citations.bib'
---



# Terrain analyses using TauDEM

This is a script that does terrain analyses using the [TauDEM tool](http://hydrology.usu.edu/taudem/taudem5/index.html) from David Tarboton at Utah State. All scripts below should be run in bash or terminal after installing TauDEM. R Code is mixed throughout to generate a final data table for analysis. 

Some R code is at the bottom of the script to project outlets and subset rasters. 

Some notes: This project evolved over the 5+ years we worked on it, and while 
I've tried to update the code to reflect my current approach, this often was too 
time-consuming or not worth it. As of posting this, all code works, assuming
you have TauDEM installed and working. Back when this was done I was pretty
bad at file tracking (still lots to learn),
so lots and lots of code chunks are set to `eval = FALSE`
to avoid rerunning time. 

Finally, I would probably do this in whitebox now since it runs with calls
inside of R, though I found TauDEM reliable and reproducible and easy to learn.

## Pit remove

TauDEM has robust documentation
```{bash, eval=F}
mpiexec -n 8 pitremove -z data/in/TauNew.tif -fel data/out/newfel.tif

mpiexec -n 8 pitremove -z data/in/TauOldElev.tif -fel data/out/oldfel.tif
```

## D8 

### Flow direction with slope and flow direction
```{bash, eval=F}
mpiexec -n 8 d8flowdir -p data/out/newdir8.tif -sd8 data/out/newsd8.tif -fel data/out/newfel.tif

mpiexec -n 8 d8flowdir -p data/out/olddir8.tif -sd8 data/out/oldsd8.tif -fel data/out/oldfel.tif

```

### Flow accumulation
```{bash, eval=F}
mpiexec -n 8 aread8 -p data/out/newdir8.tif -ad8 data/out/newuaa8.tif

mpiexec -n 8 aread8 -p data/out/olddir8.tif -ad8 data/out/olduaa8.tif
```

 
##  D-Infinity

### Flowdir
```{bash, eval=F}
mpiexec -n 8 dinfflowdir -ang data/out/newinffdr.tif -slp data/out/newinfslp.tif -fel data/out/newfel.tif

mpiexec -n 8 dinfflowdir -ang data/out/oldinffdr.tif -slp data/out/oldinfslp.tif -fel data/out/oldfel.tif


```

### Flow accumulation

```{bash, eval=F}
mpiexec -n 7 areadinf -ang data/out/newinffdr.tif -sca data/out/newinfuaa.tif

mpiexec -n 7 areadinf -ang data/out/oldinffdr.tif -sca data/out/oldinfuaa.tif

```


##  Stream detection

### Threshold 1000 pixels = 0.1 km2

```{bash, eval = F}

#New infinity
mpiexec -n 7 threshold -ssa data/out/newinfuaa.tif -src data/out/new100inf.tif -thresh 100
#Old infinity 
mpiexec -n 7 threshold -ssa data/out/oldinfuaa.tif -src data/out/old100inf.tif -thresh 100

#New D8 
mpiexec -n 7 threshold -ssa data/out/newuaa8.tif -src data/out/new1008.tif -thresh 100
#Old D8
mpiexec -n 7 threshold -ssa data/out/olduaa8.tif -src data/out/old1008.tif -thresh 100

#Curvature method



```
